[Improvement] Use system ip command instead of perl libraries to list interfaces
The libraries IO::Socket::Interfaces and IO::Interfaces::Simple returns
the list of configured network interfaces and with RUNNING flag active.
This behavior avoid that some network interfaces to be shown.

This is mainly detected in hardware appliances with several network
interfaces and debian stretch. But it could be detected in other
platforms.

Signed-off-by: Laura Garcia Liebana <lauragl@sofintel.net>
Signed-off-by: Jose Pablo <josepgl@sofintel.net>

	modified:   usr/local/zenloadbalancer/www/content3-2.cgi
	modified:   usr/local/zenloadbalancer/www/networking_functions.cgi