backend: add the new backend parameter 'ConnLimit'
If a backend has defined the 'ConnLimit' directive and it is positive,
the backend established connections will not be higher than the ConnLimit value.

The directive is set as 0 by default, this means that it is disabled and
it doesn't have any limit set.

If a new client cannot connect with any backend because they hit the limit,
the proxy will respond with a 503 error (no backend).

Signed-off-by: Alvaro Cano <alvaro.cano@zevenet.com>