http: exclude '%' from safety filter
We should assume that when a '%' character appears in a URL it is already
encoding a special character, as this is handled by the client, and as part of a
redirect it never arrives at a backend. Otherwise, it ends up re-encoding the
'%' character whenever there's a redirect, leading to "%21" in a URL, which
reference '!', becoming "%2521".