http: create tunneling state for WebSockets
The previous connection handling assumed the HTTP protocol where a request is
given and a response sent back in a synchronous order. This is not the manner in
which WebSockets works (or any tunnel for that matter). This patch enables a new
mode when tunnels are created that will freely send & receive information from
both client and backend.

Signed-off-by: Nicolás A. Ortega Froysa <nicolas.ortega@zevenet.com>