session: move JSON encoding to JSON module
The JSON encoding of sessions was temporarily found in the sessions code because
of difficulties in accessing the mutex particularly from the JSON module. Since
the refactoring of the sessions code, this is no longer an issue, and therefore
it would be a better idea to put all encoding/decoding code in the same place.

Signed-off-by: Nicolás A. Ortega Froysa <nicolas.ortega@zevenet.com>