tests: add test listing and script
Contains lists of all the tests and a script to run the different lists. The
lists are three:

 - list_pass: These work just fine as they are, and we should always ensure they
 work before submitting a new commit.
 - list_fail: These fail due to the functionality not being implemented yet.
 - list_critical: These fail because a critical error occurs, normally something
 like a segmentation fault, a gateway timeout, or something of the sort.

To run these tests simply run `./list_exec.sh <list-file>`. It will
automatically detect if the test suite has been started: if it hasn't it will
start it for you and stop it at the end of the tests.

Signed-off-by: Nicolás A. Ortega Froysa <nicolas.ortega@zevenet.com>